Decent, though possibly temporary, fix for jumping over walls vs. causing lag.
The solution was to have object._move() generate the BoundingLineSegment used in the calculation so that BoundingLineSegment.__init__() was not being called on EVERY single object. (The segment does not change at all during a call to object._move())

CollisionDetector.is_between() now takes a BoundingLineSegment instead of two points. See method signature for details.